oci-csv-json-translation/requirements.txt
oci>=2.141.1
pyyaml
pandas